Adopt commit 159d5ed from master: Merge pull request #321 from jmendelsohn4/bdlt_build_warnings-drqs-81062345
commit 159d5ed483bb07a44d217678b8802da7f65c4ca8
Merge: 2d5f24e c782215
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Thu Apr 14 13:18:27 2016 -0400

    Merge pull request #321 from jmendelsohn4/bdlt_build_warnings-drqs-81062345

    fix compiler warnings in bdlt

commit c7822156479e05a6ec58db7344a938d0230db5f4
Author: Jeffrey Mendelsohn <JMENDELSOHN4@bloomberg.net>
Date:   Thu Apr 14 11:52:40 2016 -0400

    fix compiler warnings

+++ groups/bdl/bdlt/bdlt_calendar.cpp
@@ -90,7 +90,8 @@ void Calendar::synchronizeCache()
 // PRIVATE ACCESSORS
 bool Calendar::isCacheSynchronized() const
 {
-    if (d_packedCalendar.length() != d_nonBusinessDays.length()) {
+    if (d_packedCalendar.length() !=
+                                static_cast<int>(d_nonBusinessDays.length())) {
         return false;                                                 // RETURN
     }
 
@@ -352,7 +353,8 @@ Calendar_BusinessDayConstIter::Calendar_BusinessDayConstIter(
         // Otherwise, advance the iterator to reference the next date so we can
         // find the next business day.
 
-        if (d_currentOffset == d_nonBusinessDays_p->length() - 1) {
+        if (d_currentOffset
+                      == static_cast<int>(d_nonBusinessDays_p->length()) - 1) {
             d_currentOffset = -1;
             return;                                                   // RETURN
         }
