+++ groups/bdl/bdlt/bdlt_calendar.h
@@ -1760,7 +1760,7 @@ const Date& Calendar::lastDate() const
 inline
 int Calendar::length() const
 {
-    return d_nonBusinessDays.length();
+    return static_cast<int>(d_nonBusinessDays.length());
 }
 
 inline
